+++ packages/dart_pdf_editor/lib/l10n/dart_pdf_editor_localizations_ja.dart
@@ -972,6 +972,9 @@ class DartPdfEditorLocalizationsJa extends DartPdfEditorLocalizations {
   String get shellPageGrid => 'ページグリッド';
 
   @override
+  String get shellViewPages => 'ページ';
+
+  @override
   String get shellPanelAnnotations => '注釈';
 
   @override