+++ packages/dart_pdf_editor/lib/l10n/dart_pdf_editor_localizations_ru.dart
@@ -991,6 +991,9 @@ class DartPdfEditorLocalizationsRu extends DartPdfEditorLocalizations {
   String get shellPageGrid => 'Сетка страниц';
 
   @override
+  String get shellViewPages => 'Страницы';
+
+  @override
   String get shellPanelAnnotations => 'Аннотации';
 
   @override